Make PWM resolution variable (MCU-dependent), normalise CV
When the PWM resolution is not constant, the CV range is variable, which means that gains depend on it.

To fix this, we normalise the CV everywhere from 0-1, and only multiply it by the PWM range in analogWrite()

+++ src/EEPROMConfig.h
@@ -25,7 +25,7 @@ struct ConfigStore {
 	///@}
 
 	/// Minimum solenoid duty cycle (deadband)
-	uint8_t S_lowValue;
+	double S_lowValue;
 
 	/// Total flowrate (for cascade controller) (L/min)
 	double HC_totalFlowrate;
